fix mac
fix mac

adas

fasdasdf

aaaaaaaaaaaaaaaaaaaaaahdegshjgd

Update GeometryLoader.h

Update Macos Arch reolve of lib

fix some warning issues

Update Identifiable.h

fix

+++ src/engine/rendering/RenderPass.cpp
@@ -15,7 +15,7 @@ bool RenderPass::bind(
 	if (!bindGroupIndexOpt.has_value())
 		return false;
 
-	renderPass.setBindGroup(bindGroupIndexOpt.value(), bindgroup->getBindGroup(), 0, nullptr);
+	renderPass.setBindGroup(static_cast<uint32_t>(bindGroupIndexOpt.value()), bindgroup->getBindGroup(), 0, nullptr);
 	return true;
 }
 } // namespace engine::rendering
